[Kashif]Added: Home component to display the hello from bridgelabz page using react-router-dom
