+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/reportWebVitals.js":"1","/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/App.js":"2","/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/index.js":"3"},{"size":362,"mtime":1606360849491,"results":"4","hashOfConfig":"5"},{"size":2415,"mtime":1606416672977,"results":"6","hashOfConfig":"5"},{"size":500,"mtime":1606360849441,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1l7jpag",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/reportWebVitals.js",[],"/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/App.js",[],"/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/index.js",[]]
+[{"/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/reportWebVitals.js":"1","/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/App.js":"2","/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/index.js":"3","/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/components/home/home.jsx":"4"},{"size":362,"mtime":1606360849491,"results":"5","hashOfConfig":"6"},{"size":2415,"mtime":1606416672977,"results":"7","hashOfConfig":"6"},{"size":500,"mtime":1606360849441,"results":"8","hashOfConfig":"6"},{"size":2478,"mtime":1606417292451,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1l7jpag",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/reportWebVitals.js",[],"/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/App.js",[],"/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/index.js",[],"/mnt/c/Users/aaada/NodeDevelopment/ReactDevelopment/helloworld/src/components/home/home.jsx",[]]
